Changes to secure the backend web service.

+++ src/main/java/org/hspconsortium/sandboxmanager/controllers/LaunchScenarioController.java
@@ -20,15 +20,22 @@
 
 package org.hspconsortium.sandboxmanager.controllers;
 
+import org.apache.http.HttpStatus;
 import org.hspconsortium.sandboxmanager.model.*;
 import org.hspconsortium.sandboxmanager.services.*;
 import org.springframework.web.bind.annotation.*;
 
 import javax.inject.Inject;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
 import javax.validation.Valid;
+import java.io.IOException;
 import java.io.UnsupportedEncodingException;
+import java.util.ArrayList;
+import java.util.List;
 
 @RestController
+@RequestMapping("/launchScenario")
 public class LaunchScenarioController {
 
     private final LaunchScenarioService launchScenarioService;
@@ -37,39 +44,46 @@ public class LaunchScenarioController {
     private final PersonaService personaService;
     private final AppService appService;
     private final SandboxService sandboxService;
+    private final OAuthService oAuthUserService;
 
     @Inject
     public LaunchScenarioController(final LaunchScenarioService launchScenarioService,
                                     final PatientService patientService, final PersonaService personaService,
                                     final AppService appService, final UserService userService,
-                                    final SandboxService sandboxService) {
+                                    final SandboxService sandboxService, final OAuthService oAuthUserService) {
         this.launchScenarioService = launchScenarioService;
         this.userService = userService;
         this.patientService = patientService;
         this.personaService = personaService;
         this.appService = appService;
         this.sandboxService = sandboxService;
+        this.oAuthUserService = oAuthUserService;
     }
 
-    @RequestMapping(value = "/launchScenario", method = RequestMethod.POST, consumes = "application/json", produces ="application/json")
-    public @ResponseBody LaunchScenario createLaunchScenario(@RequestBody @Valid final LaunchScenario launchScenario) {
-        User user = userService.findByLdapId(launchScenario.getOwner().getLdapId());
-        if (user == null) {
-            user = userService.save(launchScenario.getOwner());
-        }
-        launchScenario.setOwner(user);
-
+    @RequestMapping(method = RequestMethod.POST, consumes = "application/json", produces ="application/json")
+    public @ResponseBody LaunchScenario createLaunchScenario(HttpServletRequest request, @RequestBody @Valid final LaunchScenario launchScenario) {
         // A null sandbox is the HSPC sandbox
         Sandbox sandbox = null;
         if (launchScenario.getSandbox() != null) {
             sandbox = sandboxService.findBySandboxId(launchScenario.getSandbox().getSandboxId());
+            checkUserAuthorization(request, sandbox.getCreatedBy().getLdapId());
             launchScenario.setSandbox(sandbox);
         }
+
+        checkUserAuthorization(request, launchScenario.getCreatedBy().getLdapId());
+        User user = userService.findByLdapId(launchScenario.getCreatedBy().getLdapId());
+        if (user == null) {
+            user = userService.save(launchScenario.getCreatedBy());
+        }
+        launchScenario.setCreatedBy(user);
+        List<User> users = new ArrayList<>(1);
+        users.add(user);
+        launchScenario.setUsers(users);
 
         Persona persona = null;
         if (launchScenario.getPersona() != null) {
             if (sandbox == null) {
-                persona = personaService.findByFhirId(launchScenario.getPersona().getFhirId());
+                persona = personaService.findByFhirIdAndSandboxId(launchScenario.getPersona().getFhirId(), null);
             } else {
                 persona = personaService.findByFhirIdAndSandboxId(launchScenario.getPersona().getFhirId(), sandbox.getSandboxId());
             }
@@ -84,7 +98,7 @@ public class LaunchScenarioController {
         if (launchScenario.getPatient() != null) {
             Patient patient = null;
             if (sandbox == null) {
-                patient = patientService.findByFhirId(launchScenario.getPatient().getFhirId());
+                patient = patientService.findByFhirIdAndSandboxId(launchScenario.getPatient().getFhirId(), null);
             } else {
                 patient = patientService.findByFhirIdAndSandboxId(launchScenario.getPatient().getFhirId(), sandbox.getSandboxId());
             }
@@ -98,7 +112,7 @@ public class LaunchScenarioController {
 
         App app = null;
         if (sandbox == null) {
-            app = appService.findByClientId(launchScenario.getApp().getClient_id());
+            app = appService.findByClientIdAndSandboxId(launchScenario.getApp().getClient_id(), null);
         } else {
             app = appService.findByClientIdAndSandboxId(launchScenario.getApp().getClient_id(), sandbox.getSandboxId());
         }
@@ -112,34 +126,59 @@ public class LaunchScenarioController {
         return launchScenarioService.save(launchScenario);
     }
 
-    @RequestMapping(value = "/launchScenario", method = RequestMethod.PUT, produces ="application/json")
-    public @ResponseBody LaunchScenario updateLaunchScenario(@RequestBody @Valid final LaunchScenario launchScenario) {
+    @RequestMapping(method = RequestMethod.PUT, produces ="application/json")
+    public @ResponseBody LaunchScenario updateLaunchScenario(HttpServletRequest request, @RequestBody @Valid final LaunchScenario launchScenario) {
+        checkUserAuthorization(request, launchScenario.getCreatedBy().getLdapId());
         LaunchScenario updateLaunchScenario = launchScenarioService.getById(launchScenario.getId());
-        updateLaunchScenario.setLastLaunchSeconds(launchScenario.getLastLaunchSeconds());
-        return launchScenarioService.save(updateLaunchScenario);
+        if (updateLaunchScenario != null) {
+            updateLaunchScenario.setLastLaunchSeconds(launchScenario.getLastLaunchSeconds());
+            return launchScenarioService.save(updateLaunchScenario);
+        }
+        return null;
     }
 
-    @RequestMapping(value = "/launchScenario", method = RequestMethod.DELETE, produces ="application/json")
-    public @ResponseBody void deleteLaunchScenario(@RequestBody @Valid final LaunchScenario launchScenario) {
+    @RequestMapping(method = RequestMethod.DELETE, produces ="application/json")
+    public @ResponseBody void deleteLaunchScenario(HttpServletRequest request, @RequestBody @Valid final LaunchScenario launchScenario) {
+        checkUserAuthorization(request, launchScenario.getCreatedBy().getLdapId());
         launchScenarioService.delete(launchScenario);
     }
 
-    @RequestMapping(value = "/launchScenarios", method = RequestMethod.GET, produces ="application/json",
-            params = {"id", "sandboxId"})
-    public @ResponseBody Iterable<LaunchScenario> getLaunchScenarios(@RequestParam(value = "id") String id, @RequestParam(value = "sandboxId") String sandboxId) {
-        String ownerId = null;
+    @RequestMapping(method = RequestMethod.GET, produces ="application/json",
+            params = {"userId", "sandboxId"})
+    public @ResponseBody Iterable<LaunchScenario> getLaunchScenarios(HttpServletRequest request,
+        @RequestParam(value = "userId") String userIdEncoded, @RequestParam(value = "sandboxId") String sandboxId) throws UnsupportedEncodingException{
 
-        try {
-            ownerId = java.net.URLDecoder.decode(id, "UTF-8");
-            // A null sandbox is the HSPC sandbox
-            if (sandboxId == null || sandboxId.isEmpty()) {
-                return launchScenarioService.findByOwnerId(ownerId);
-            } else {
-                return launchScenarioService.findByOwnerIdAndSandboxId(ownerId, sandboxId);
-            }
-        } catch (UnsupportedEncodingException e) {
-            e.printStackTrace();
+        String userId = java.net.URLDecoder.decode(userIdEncoded, "UTF-8");
+        checkUserAuthorization(request, userId);
+        // A null sandbox is the HSPC sandbox
+        if (sandboxId == null || sandboxId.isEmpty()) {
+            return launchScenarioService.findByUserIdAndSandboxId(userId, null);
+        } else {
+            return launchScenarioService.findByUserIdAndSandboxId(userId, sandboxId);
         }
-        return null;
+    }
+
+    @ExceptionHandler(UnauthorizedException.class)
+    @ResponseBody
+    @ResponseStatus(code = org.springframework.http.HttpStatus.UNAUTHORIZED)
+    public void handleAuthorizationException(HttpServletResponse response, Exception e) throws IOException {
+        response.getWriter().write(e.getMessage());
+    }
+
+    @ExceptionHandler(Exception.class)
+    @ResponseBody
+    @ResponseStatus(org.springframework.http.HttpStatus.BAD_REQUEST)
+    public void handleException(HttpServletResponse response, Exception e) throws IOException {
+        response.getWriter().write(e.getMessage());
+    }
+
+    private void checkUserAuthorization(HttpServletRequest request, String userId) {
+        String oauthUserId = oAuthUserService.getOAuthUserId(request);
+
+        if (!userId.equalsIgnoreCase(oauthUserId)) {
+            throw new UnauthorizedException(String.format("Response Status : %s.\n" +
+                    "Response Detail : User not authorized to perform this action."
+                    , HttpStatus.SC_UNAUTHORIZED));
+        }
     }
 }
